Add 'create' param to #create and #initialize
Status#create and Status#initialize now support an optional create
parameter.  If it is given and is false and the status buffer for the
given instance ID does not exist, then an Errno:ENOENT exception is
raised.  Otherwise, the status buffer for the given instance is created
if it dos not already exist.  The equivalent ruby declarations for these
methods are:

def initialize(instance_id, create=true)
def attach(instance_id, create=true)