Put calls on separate lines to make debug easier
This splits a "foo(bar())" construct into two separate lines:

b=bar()
foo(b)

to make debugging easier.  Now if a thread gets "stuck" in this section,
it will be easier to tell if it is stuck in bar() or in foo().  See the
diff for the actual functions involved (foo() and bar() are not the real
names).