Add new pubsub option to hashpipe_redis_gateway.rb
hashpipe_redis_gateway.rb now subscribes to a global and instance
specific request channel.  The global request channel is
`hashpipe:///req`.  The instance specific request channel is
`hashpipe://GW_NAME/INST_ID/req` (e.g. `hashpipe://blc00/0/req`).

The request channels are used to request values from the gateway's
status buffer.  The format of the message is a single keyword or a
newline ("\n") delimited list of keywords whose values are being
requested.

Upon receipt of a request channel message, the gateway will publish a
reply message on an instance specific reply channel.  The instance
specific reply channel is `hashpipe://GW_NAME/INST_ID/rep` (e.g.
`hashpipe://blc00/0/rep`).  The format of the reply message is a list of
KEYWORD=VALUE pairs separated by newlines, one pair for each keyword
requested.  Typically the publisher of the request message will
subscribe to the reply channel so that it will receive the reply
messages, but this is not strictly necessary (maybe the requestor is
requesting on behalf of some other process that has subscribed to the
reply channel).