Give HRG the ability to join/leave groups
hashpipe_redis_gateway.rb now supports the notion of "groups".
Publishing a "join" message to the command channel of an instance (or
gateway or broadcast) will cause the instance (or all instances) to join
the group.  If any instance has joined a group, the gateway will
subscribe to a group "set" channel: "<domain>:<group>///set".  The
gateway keeps track of which of its instances have joined a given group
so that messages published to the group "set" channel will only affect
instances of that gateway that have actually joined the group.

A corresponding "leave" message can be published to cause instances to
leave a group.  When no more instances belong to a group, the gateway
unsubscribes from the corresponding group "set" channel.